create SupervisedJobTask
My first attempt at rewriting Custodian with a number of changes and also making it a Prefect Task. Plenty of notes are included inside.  This is also not robustly tested but with an initial run it looks like the Prefect import are slow but this class is >10x faster than Custodian otherwise.