remove regrid and refinement from electride finder
Electrides were being placed at odd spots. Regridding seemed to contribute to this and the method I had for refining didn't seem to help. Its possible that I'll need to put a refinement method back in later, but that will depend on if it seems to effect anything in structures I havent tested yet.