Tweeks to BadELF
This commit makes changes to fix a number of issues found when testing BadELF on ~50 electrides. The main change is the removal of one of the partitioning plane reduction algorithms which would sometimes remove important planes. An alternative version will need to be added in the future to reduce computation times. Other changes include the addition of a minimum electride radius in the electride_finder module and two new columns in the badelf Model.