Switch to Grid based Dimensionality
The previous dimensionality algorithm relied on connections between electride sites and the pymatgen get_dimensionality_larsen method. There were several issues where this method failed, some of which I haven't figured out. One fundamental issue was that the ELF can connect at a given isosurface level in ways that are not along the bonds between electride sites. The new method allows for these connections and is faster, but has the slight disadvantage that the connecting values are not interpolated and are therefore dependent on the voxel grid density. However, this is probably less than the error inherent to the ELF.