Update get_elf_radius and covalency check
The _get_elf_ionic_radius method is primarily used in the electride finder. In some electrides there are sodium atoms surrounded by ELF which can be very far from atoms. There was already a distance check for these, but for one electride (Ba3na4N) this didn't work. I've added a check for if the bond is covalent.  I'm also testing a new method for checking structures for covalency and I've updated the covalency check accordingly.