Test fix for partitioning bug
With the updated partitioning scheme, there are some differences in where partitioning planes are located. This is most apparent in VoronELF. The main issue here was that I assigned the partitioning fraction based on the closest maximum to the center if voxels assigned to atoms unrelated to the bond were found in the partitioning line. If there was no maximum, this resulted in fractions very close to atom centers. This has been corrected. Another potential problem is if there is a maximum that is assigned to one of the atoms in the bond, but the maximum actually comes from a symmetrically equivalent but distinct atom. This would require significantly more work I think to fix so I'm leaving it for now.