Add reduced partitioning
For the initial voxel assignment, the reduced planes aren't enough (at least in their current version). However, the large number of planes before reduction result in far too many planes to be able to determine which are relavent for the remaining voxel assignment. To deal with this I'm using the full partitioning for the first assignment and the reduced partitioning for the second part. It's possible this may cause issues if one of the removed planes is in fact important, but I think this is the best bet for now.