update multi-site atom distance
In some cases, the cutoff for the multi-site atom assignment was too small. This happened in the Hf2X type electrides. The radius to consider the X atom ends up being their elf ionic radii which results in a thin line connecting electride sites to make it 3D. Increasing the radius to 1.2*(furthest_neigh_dist-neigh_radii) fixes this issue.