Heavy Change to Unassigned Voxels
Assigning voxels to the nearest plane seems to be impossible or very difficult/computationally expensive as you would need to know the extent of the plane and calculate the distance from each voxel to the plane segment. Instead I've switched to a different method that relies on indexing rather than mathematical operations. The assignments of the voxels in a small sphere around the target voxels are found and the ratio of these assignments is found. If no nearby assignments are found, the sphere is made slightly larger and the operation repeats. This appears to give equal charges while also assigning the remaining voxels as I would expect.