Add StagedCalculation Model
To make things easier on users, I've added a StagedCalculation model that should be used by any staged workflows. The model inherits the Calculation model and adds columns for subworkflow names and ids. It also has convenience methods to obtain the subworkflow results from their respective tables.

In addition to this, I've made it so that the fixed_composition evo search looks to see if the selected workflow is a staged calculation. If it is, it uses the database of the last workflow of the staged calculation when filtering for finished calculations. This way the only things the user needs to do for custom calculations is define subworkflows and make sure the final workflow returns the desired fitness field. For custom fitness fields they will need to make a custom model, where the only requirement is that it inherits the Structure model.