Add Error Check to Staged Workflow
The staged workflow and corresponding model now check if the workflows fail and if they do, record which one.

Additionally, necessary arguments for subworkflows can be passes through the run argument as a dictionary "subworkflow_kwargs". So the only requirement now is that a Structure object is passed.

Warren lab and badelf apps were updated to use this new staged workflow setup.