Adjust Steadystate Source Submission
Steadystate sources submitted up to an allowed integer. This is problematic as if one source tends to result in structures that are easier to relax than another, it will submit more than its ideal proportion of structures.

The submission now looks at all submissions and submits from the source that is most needed to reach the desired proportion.

This will help with the next step of dynamically updating sources based on various criteria.

There are still some bugs to work on.